requirements.txt
Django==4.2.5
gunicorn==21.2.0
whitenoise==6.5.0
Pillow==10.0.0
